djngular/requirements.txt
amqp==5.0.9
asgiref==3.5.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
celery-progress==0.1.2
cffi==1.15.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
Django==4.0.2
django-celery-results==2.2.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
dnspython==2.2.0
eventlet==0.33.0
gevent==21.12.0
greenlet==1.1.2
kombu==5.2.3
packaging==21.3
prompt-toolkit==3.0.28
pycparser==2.21
pyparsing==3.0.7
pytz==2021.3
redis==3.2.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zope.event==4.5.0
zope.interface==5.4.0
